Interval and subclasses now use ints rather than longs internally. This might make a big diff in memory consumption, as we have 1 Interval / retransmission task

+++ src/org/jgroups/stack/ExponentialInterval.java
@@ -4,14 +4,14 @@ package org.jgroups.stack;
  * @author Bela Ban
  */
 public class ExponentialInterval implements Interval {
-    private long value=30;
-    private static final long MAX=15000;
+    private int value=30;
+    private static final int MAX=15000;
 
     public ExponentialInterval() {
         
     }
 
-    public ExponentialInterval(long value) {
+    public ExponentialInterval(int value) {
         this.value=value;
     }
